Moving away from a global 'bindings' namespace
