+++ src/zerodbext/aead/bindings/__init__.py
@@ -14,13 +14,13 @@
 
 from __future__ import absolute_import, division, print_function
 
-from zerodbext.bindings.crypto_aead_aes256gcm import (
+from .crypto_aead_aes256gcm import (
     crypto_aead_aes256gcm_ABYTES, crypto_aead_aes256gcm_KEYBYTES,
     crypto_aead_aes256gcm_NPUBBYTES, crypto_aead_aes256gcm_decrypt,
     crypto_aead_aes256gcm_encrypt, crypto_aead_aes256gcm_is_available,
 )
 
-from zerodbext.bindings.sodium_core import sodium_init
+from .sodium_core import sodium_init
 
 
 __all__ = [
