+++ tests/test_aes256gcm.py
@@ -1,4 +1,3 @@
-
 from __future__ import absolute_import, division, print_function
 
 import binascii
